Add initial display implementation
This adds the initial implementation for writing to the keyboard
displaying. Instead writing each "image" line by line (including the
ones we've already written) we keep two buffers, one what is currently
displayed on the keyboard, and one that is the new value. We
continuously loop over the image and only update the keys that changed