Get basic Display implementation working
This adds the basic ability to create a running display loop, and send
updates asynchronously to get drawn on the next display update loop.
Right now this uses a whole bunch of struct passing, and this is due to
the many pointer issues I ran into during my original implementation.
This can/should be changed later, but for now this works and therefore
is ideal