Add back checking if we are animating when updating the display
I had this before and got rid of it, but it is actually pretty
important. We need to know if we are running an animation loop in a
separate go channel so that we can send it a stop signal. With this
change, we can go back and forth between displaying the canada flag and
doing the fish animation. Its pretty dope